[NFC] Static cast optional bit-fields
llvm::Optional had a constructor that took a const l-value
reference, but std::optional does not, resulting in build failures when
returning bit-fields in optionals. Casting the bit-field to the
appropriate type fixes this.

+++ lib/ClangImporter/ImportName.cpp
@@ -2378,7 +2378,7 @@ ImportedName NameImporter::importNameImpl(const clang::NamedDecl *D,
           method->getDeclContext(), getNonNullArgs(method, params),
           result.getErrorInfo()
               ? llvm::Optional<unsigned>(
-                    result.getErrorInfo()->ErrorParameterIndex)
+                    static_cast<unsigned int>(result.getErrorInfo()->ErrorParameterIndex))
               : llvm::None,
           method->hasRelatedResultType(), method->isInstanceMethod(),
           result.getAsyncInfo().transform(
